More entity plugin cleanup
1. Separate plugins into their own modules
2. Split up dream spinners
3. Remove corner boost block placements
